Fix email validation, security bugs, and frontend API contract mismatches
- Email regex now accepts uppercase letters, multiple separators, and
  hyphens in the local part instead of rejecting most valid addresses
- Fix path traversal and duplicate-filename overwrite on upload by
  storing files under a UUID instead of the client-supplied name
- Enforce storage quota on upload and track used_storage automatically
  (used_storage changed from Integer to Float, fractional GB was
  truncated to 0 before)
- Sanitize file/folder names against path traversal and zip-slip, and
  make response headers Latin-1 safe to avoid crashing on unicode names
- Fix null nonce crashing download headers and zip export
- Align /user/updatelogin and /user/updateusedstorage routes/bodies
  with what the WPF frontend actually sends
- Replace HTTPException-as-success-response anti-pattern with proper
  return values, add missing /user/logout endpoint

+++ Backend/models.py
@@ -11,7 +11,10 @@ class DBUser(Base):
     email = Column(String)
     password = Column(String)
     last_login = Column(DateTime)
-    used_storage = Column(Integer, default=0)   # in Gigabytes
+    # AI Agent: war Integer - Dateigroessen in GB sind fast immer
+    # Bruchzahlen (z.B. 0.003 GB), mit Integer wurde jede Quota-Buchung
+    # auf 0 abgeschnitten und das Limit damit nie erreicht.
+    used_storage = Column(Float, default=0)     # in Gigabytes
     storage_plan = Column(Integer)              # in Gigabytes
